Basic Simulation
When the room numbers are input, the code then prints the trolley actions accordingly. It shows all the motor actions the trolley undergoes to reach a room and also takes into consideration the case if a non existent room is entered. 

Nb : The layout of the floor is assumed to be a rectangular floor with rooms on either side of corridor. There are 8 rooms, 1,3,5,7 on the right side of the corridor and the rest to the left. 1 and 2 are the nearest rooms and 7 and 8 the farthest.